+++ validatorController.js
@@ -1,7 +1,6 @@
 const express = require("express")
 const mailValidator = require("email-validator");
 const validarCpf = require('validar-cpf');
-const Sequelize=require("sequelize")
 const Request = require("request-promise-native");
 /**
 * @param {Array} validationRoles
@@ -23,16 +22,28 @@ module.exports.validate=async (validationRoles,obj)=>{
         valida=await typeValidation(vlg,body)
         if(valida)
             retorno.push(valida)
-        //UNIQUE
-        valida=await uniqueValidation(vlg,body)
-        if(valida)
-            retorno.push(valida)
         //MINLENGTH
         valida=minLengthValidation(vlg,body)
         if(valida)
             retorno.push(valida)
-        //EXISTS
-        valida=await existsValidation(vlg,body)
+        //LENGTH
+        valida=lengthValidation(vlg,body)
+        if(valida)
+            retorno.push(valida)
+        //MAXLENGTH
+        valida=maxLengthValidation(vlg,body)
+        if(valida)
+            retorno.push(valida)
+        //MIN
+        valida=minValidation(vlg,body)
+        if(valida)
+            retorno.push(valida)
+        //MAX
+        valida=maxValidation(vlg,body)
+        if(valida)
+            retorno.push(valida)
+        //EQUAL
+        valida=equalValidation(vlg,body)
         if(valida)
             retorno.push(valida)
     }
@@ -87,49 +98,46 @@ async function typeValidation(validationRole,body){
     return validationRole.type[1]
 }
 
-async function existsValidation(validationRole,body){
-    if(!validationRole.exists || !body[validationRole.param])
-        return
-    const env=process.env
-    const sequelize=new Sequelize(env.DATABASE,env.USERDB,env.PASSWORDDB,{
-            host:env.HOSTDB,
-            dialect:env.DIALECTDB
-        }
-    )
-    const queryParams=validationRole.exists[0].split(":")
-    let query="select count(*) as count from "+queryParams[0]+" where "+validationRole.exists[1]+" = '"+body[validationRole.param]+"'"
-    const count=await sequelize.query(query,{type:Sequelize.QueryTypes.SELECT})
-    sequelize.close()
-    if(count[0]["count"]<1)
-        return validationRole.exists[2]
-    return
-}
-
-async function uniqueValidation (validationRole,body){
-    if(!validationRole.unique || !body[validationRole.param])
-        return
-    const env=process.env
-    const sequelize=new Sequelize(env.DATABASE,env.USERDB,env.PASSWORDDB,{
-            host:env.HOSTDB,
-            dialect:env.DIALECTDB
-        }
-    )
-    const queryParams=validationRole.unique[0].split(":")
-    let query="select count(*) as count from "+queryParams[0]+" where "+validationRole.unique[1]+" = '"+body[validationRole.param]+"'"
-    if(queryParams.length>1)
-        query+=" and "+queryParams[1]+" != "+queryParams[2]
-    const count=await sequelize.query(query,{type:Sequelize.QueryTypes.SELECT})
-    sequelize.close()
-    if(count[0]["count"]>0)
-        return validationRole.unique[2]
-    return
-}
-
 function minLengthValidation(validationRole,body){
     if(!validationRole.minLength || !body[validationRole.param])
         return
     if(body[validationRole.param].length<validationRole.minLength[0])
         return validationRole.minLength[1]
+    return
+}
+function maxLengthValidation(validationRole,body){
+    if(!validationRole.maxLength || !body[validationRole.param])
+        return
+    if(body[validationRole.param].length<validationRole.maxLength[0])
+        return validationRole.maxLength[1]
+    return
+}
+function lengthValidation(validationRole,body){
+    if(!validationRole.length || !body[validationRole.param])
+        return
+    if(body[validationRole.param].length<validationRole.length[0])
+        return validationRole.length[1]
+    return
+}
+function minValidation(validationRole,body){
+    if(!validationRole.min || !body[validationRole.param])
+        return
+    if(body[validationRole.param]<validationRole.min[0])
+        return validationRole.min[1]
+    return
+}
+function maxValidation(validationRole,body){
+    if(!validationRole.max || !body[validationRole.param])
+        return
+    if(body[validationRole.param]>validationRole.max[0])
+        return validationRole.max[1]
+    return
+}
+function equalValidation(validationRole,body){
+    if(!validationRole.equal || !body[validationRole.param])
+        return
+    if(body[validationRole.param]===validationRole.equal[0])
+        return validationRole.equal[1]
     return
 }