+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+---
+
+## [0.0.3] - yyyy.mm.dd
+### Added
+- types
+    - array
+- forbiden
+
+### Changed
+- exists
+    - Now exists can validade arrays.
+### Debuged
+- type number validation
+---
+
 ## [0.0.2] - 2019-08-13
 ### Added
 - Dependences